v1: Add generic single threaded tcp server
server.hpp introduces a server and a connection pool. The generic classes in
there could be used to specify any specific way of serving a tcp connection by
providing a coroutine. Everything runs on one thread, by leveraging C++20
coroutines.

One example of a tcp echo-server implemented using the generic classes in this
library can be seen in main.cpp.